+++ Classes/OldModel/Registration.php
@@ -460,10 +460,10 @@ class Tx_Seminars_OldModel_Registration extends \Tx_Seminars_OldModel_Abstract
                 // The fallthrough is intended.
             case 'tstamp':
                 $format = $this->getConfValueString('dateFormatYMD') . ' ' . $this->getConfValueString('timeFormat');
-                $result = \strftime($format, $rawData);
+                $result = \strftime($format, (int)$rawData);
                 break;
             case 'date_of_birth':
-                $result = \strftime($this->getConfValueString('dateFormatYMD'), $rawData);
+                $result = \strftime($this->getConfValueString('dateFormatYMD'), (int)$rawData);
                 break;
             case 'name':
                 $result = $this->getFrontEndUser()->getName();